Change AppSettings -> AppSettingsManager
- Rename AddSettings -> AppSettingsManager (standard naming)
- Create AppSettings.cs file that is a serializable class (for JSON)
