+++ MBBSEmu/HostProcess/MbbsHost.cs
@@ -86,7 +86,7 @@ namespace MBBSEmu.HostProcess
         /// <summary>
         ///     Configuration Class giving access to the appsettings.json file
         /// </summary>
-        private readonly AppSettings _configuration;
+        private readonly AppSettingsManager _configuration;
 
         /// <summary>
         ///     Time of day that the cleanup routine will trigger
@@ -119,7 +119,7 @@ namespace MBBSEmu.HostProcess
         private readonly IAccountRepository _accountRepository;
         private readonly ITextVariableService _textVariableService;
 
-        public MbbsHost(IClock clock, ILogger logger, IGlobalCache globalCache, IFileUtility fileUtility, IEnumerable<IHostRoutine> mbbsRoutines, AppSettings configuration, IEnumerable<IGlobalRoutine> globalRoutines, IAccountKeyRepository accountKeyRepository, IAccountRepository accountRepository, PointerDictionary<SessionBase> channelDictionary, ITextVariableService textVariableService, IMessagingCenter messagingCenter)
+        public MbbsHost(IClock clock, ILogger logger, IGlobalCache globalCache, IFileUtility fileUtility, IEnumerable<IHostRoutine> mbbsRoutines, AppSettingsManager configuration, IEnumerable<IGlobalRoutine> globalRoutines, IAccountKeyRepository accountKeyRepository, IAccountRepository accountRepository, PointerDictionary<SessionBase> channelDictionary, ITextVariableService textVariableService, IMessagingCenter messagingCenter)
         {
             Logger = logger;
             Clock = clock;
